Move phrase usages within the phrase controller
