+++ app/controllers/phrases_controller.rb
@@ -1,10 +1,15 @@
-class PhraseController < ApplicationController
+class PhrasesController < ApplicationController
   def show
     @phrase = Phrase.find(params[:id])
     @devices = devices
     @pages_visited = pages_visited
     @mentions = mentions
     @survey_answers_containing_phrase = SurveyAnswer.find_by_sql(["select * from survey_answers sa join questions q on q.id = sa.question_id join survey_phrases sp on sp.survey_answer_id = sa.id join phrases p on p.id = sp.phrase_id where p.id = ? and q.question_number = 3 and sa.answer not like '-' limit 3", "#{@phrase.id}"])
+  end
+
+  def usage
+    @phrase = Phrase.find(params[:id])
+    @presenter = PhraseUsagePresenter.new(survey_answers_containing_phrase, search_params)
   end
 
 private
@@ -28,5 +33,19 @@ private
       .mentions_by_date_range_for_phrase(@phrase, Date.new(2020, 4, 1), Date.new(2020, 4, 7))
       .sort_by{ |date, _| date }
       .to_h
+  end
+
+  def survey_answers_containing_phrase
+    SurveyAnswer.find_by_sql(["select * from survey_answers sa join questions q on q.id = sa.question_id join survey_phrases sp on sp.survey_answer_id = sa.id join phrases p on p.id = sp.phrase_id where p.id = ? and q.question_number = 3 and sa.answer not like '-'", "#{params[:id]}"])
+  end
+
+  def search_params
+    @search_params ||= begin
+      { page: 1 }.merge(
+        params.permit(
+          :page,
+        ).to_h.symbolize_keys
+      )
+    end
   end
 end
