+++ app/helpers/phrase_helper.rb
@@ -7,4 +7,8 @@ module PhraseHelper
       ]
     end
   end
+
+  def highlighted_survey_answer_html(survey_answer, phrase)
+    survey_answer.gsub(/(#{phrase})/i){ |match| "<span class='phrase-highlight'>#{match}</span>" }.html_safe
+  end
 end
