--- spec/controllers/phrase_usage_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe PhraseUsageController, type: :controller do
-
-end
